Module Paccaml.History

val init : unit -> int * int

init () initialize the history file. If there is history of playing, load it in. Otherwise, make the file. first int is lifetime points, second is highscore

val save : int -> unit

save points takes a number of points and updates history as a result. First element of history is the lifetime points. second is the high score