Module Paccaml.Mapbuilder

val main : int -> int -> unit

main x y runs the series of code that allows a user to make and save a valid map using a graphic interface (maps are valid if there is a ghoststart tile, hometile, and all tiles are reachable)